Improved serial communication, hb-unit and r-unit now communication successfully, implemented handshake function

+++ Ctrl-MC_RelayUnit/src/mainInit.h
@@ -7,7 +7,7 @@
 class MainInit {
     // Set Arduino pins according to config
     private:
-        void setRelayOutputPin(int pin, bool defaultOn) {
+        void setRelayOutputPin(uint8_t pin, bool defaultOn) {
             if (pin > -1)
             {
                 pinMode(pin, OUTPUT);
